refactor file names to kebab style
