+++ src/repository/index.js
@@ -1,2 +1,2 @@
-export * from './user.repository.js';
-export * from './category.repository.js';
+export * from './user-repository.js';
+export * from './category-repository.js';
