+++ src/routes/index.js
@@ -1,2 +1,2 @@
-export { default as userRouter } from './user.routes.js';
-export { default as categoryRouter } from './category.routes.js';
+export { default as userRouter } from './user-routes.js';
+export { default as categoryRouter } from './category-routes.js';
