+++ src/service/index.js
@@ -1,2 +1,2 @@
-export * from './user.service.js';
-export * from './category.service.js';
+export * from './user-service.js';
+export * from './category-service.js';
